Move asdf trace parsing to a separate module
