+++ sourcespec2/input/trace_parsers/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf8 -*-
+# SPDX-License-Identifier: CECILL-2.1
+"""
+Trace parsers for SourceSpec.
+
+:copyright:
+    2013-2024 Claudio Satriano <satriano@ipgp.fr>
+:license:
+    CeCILL Free Software License Agreement v2.1
+    (http://www.cecill.info/licences.en.html)
+"""
+from .asdf_traces import parse_asdf_traces  # noqa
