+++ sourcespec2/input/traces.py
@@ -21,11 +21,11 @@ import shutil
 import tarfile
 import zipfile
 import tempfile
-import json
 from obspy import read
 from obspy.core import Stream
 from ..setup import config, ssp_exit
 from .sac_header import get_event_from_SAC
+from .trace_parsers import parse_asdf_traces
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
 
 
@@ -70,126 +70,46 @@ def _build_filelist(path, filelist, tmpdir):
             filelist.append(path)
 
 
-def _parse_asdf_trace_headers(ds, st, nw_stat_codes, tag):
+def _filter_by_station(input_stream):
     """
-    Parse ASDF trace headers.
+    Select traces for a given station, if specified in the configuration.
 
-    :param ds: ASDF dataset
-    :type ds: :class:`pyasdf.ASDFDataSet`
-    :param st: ObsPy Stream object
-    :type st: :class:`obspy.core.stream.Stream`
-    :param nw_stat_codes: Network and station codes
-    :type nw_stat_codes: list
-    :param tag: waveform tag in ASDF file
-    :type tag: str
+    :param input_stream: Input stream to filter
+    :type tmpst: :class:`obspy.core.stream.Stream`
+
+    :return: Stream object
+    :rtype: :class:`obspy.core.stream.Stream`
     """
-    # pylint: disable=import-outside-toplevel
-    from pyasdf.utils import AuxiliaryDataContainer
-    for nw_stat_code, tr in zip(nw_stat_codes, st.traces):
-        if nw_stat_code in ds.auxiliary_data['TraceHeaders']:
-            auxiliary_root = ds.auxiliary_data['TraceHeaders'][nw_stat_code]
-        else:
-            # ESM
-            _nw_stat_code = nw_stat_code.replace('.', '_')
-            auxiliary_root = ds.auxiliary_data['TraceHeaders'][_nw_stat_code]
-        if not tag or tag not in auxiliary_root:
-            continue
-        header = None
-        tr_id = tr.id.replace('.', '_')
-        if isinstance(auxiliary_root[tag], AuxiliaryDataContainer):
-            # ESM
-            header = auxiliary_root[tag].parameters
-        elif tr_id in auxiliary_root[tag]:
-            header = auxiliary_root[tag][tr_id].parameters
-        if not header:
-            continue
-        for key, val in header.items():
-            try:
-                val = json.loads(val)
-            except json.JSONDecodeError:
-                if isinstance(val, type(b'')):
-                    # ESM
-                    val = val.decode('ascii')
-                if key == 'processing':
-                    val = [val]
-            # Try preserving original _format
-            if (
-                key == '_format'
-                and val != 'ASDF'
-                and 'original_format' not in header.items()
-            ):
-                key = 'original_format'
-            # Write to trace header
-            if key not in tr.stats:
-                tr.stats[key] = val
+    if config.options.station is None:
+        return input_stream
+    return Stream([
+        trace for trace in input_stream.traces
+        if trace.stats.station == config.options.station
+    ])
 
 
-def parse_asdf_traces(asdf_file, tag=None, read_headers=False):
+def _read_asdf_traces():
     """
-    Read all traces from ASDF file with given waveform tag
-    If tag is not specified, the first available one will be taken
-
-    :param asdf_file: full path to ASDF file
-    :type asdf_file: str
-    :param tag: waveform tag in ASDF file
-    :type tag: str
-    :param read_headers: flag to control reading of (non-standard)
-        trace headers
-    :type read_headers: bool
+    Read traces from ASDF file specified in the configuration.
 
     :return: ObsPy Stream object
     :rtype: :class:`obspy.core.stream.Stream`
     """
-    try:
-        # pylint: disable=import-outside-toplevel
-        # pyasdf is not a hard dependency, so we import it here
-        # and check for ImportError
-        import pyasdf
-    except ImportError:
-        logger.error(
-            'Error importing pyasdf. '
-            'See https://seismicdata.github.io/pyasdf/ for installation '
-            'instructions.'
-        )
-        ssp_exit(1)
-    try:
-        ds = pyasdf.ASDFDataSet(asdf_file, mode='r')
-    except Exception as err:
-        logger.error(err)
-        ssp_exit(1)
-    # Read waveform data
-    st = Stream()
-    nw_stat_codes = []
-    for nw_stat_code in ds.waveforms.list():
-        wf_tags = ds.waveforms[nw_stat_code].get_waveform_tags()
-        # If tag is not specified, take first available tag
-        if not tag:
-            # Maybe this should be logged
-            tag = wf_tags[0]
-        if tag in wf_tags:
-            station_st = ds.waveforms[nw_stat_code][tag]
-            st.extend(station_st)
-            nw_stat_codes.extend([nw_stat_code] * len(station_st))
-    # Try reading trace headers if present
-    if 'TraceHeaders' in ds.auxiliary_data:
-        header_key = 'TraceHeaders'
-    elif 'Headers' in ds.auxiliary_data:
-        header_key = 'Headers'
-    else:
-        header_key = None
-    if read_headers and header_key:
-        _parse_asdf_trace_headers(ds, st, nw_stat_codes, tag)
-    ds._close()
-    return st
+    asdf_file = config.options.asdf_file
+    if not asdf_file:
+        return Stream()
+    return _filter_by_station(parse_asdf_traces(asdf_file))
 
 
 def _read_trace_files():
     """
-    Read trace files from the path specified in the configuration file.
+    Read trace files from the path specified in the configuration.
 
     :return: ObsPy Stream object
     :rtype: :class:`obspy.core.stream.Stream`
     """
+    if config.options.trace_path is None:
+        return Stream()
     # phase 1: build a file list
     # ph 1.1: create a temporary dir and run '_build_filelist()'
     #         to move files to it and extract all tar archives
@@ -207,21 +127,11 @@ def _read_trace_files():
     st = Stream()
     for filename in sorted(filelist):
         try:
-            if os.path.splitext(filename)[-1].lower() in ('.asdf', '.h5'):
-                tmpst = parse_asdf_traces(filename)
-            else:
-                tmpst = read(filename, fsize=False)
-        except Exception:
+            st += _filter_by_station(read(filename, fsize=False))
+        except (TypeError, FileNotFoundError):
             logger.warning(
                 f'{filename}: Unable to read file as a trace: skipping')
             continue
-        for trace in tmpst.traces:
-            # only use the station specified by the command line option
-            # "--station", if any
-            if (config.options.station is not None and
-                    trace.stats.station != config.options.station):
-                continue
-            st.append(trace)
     shutil.rmtree(tmpdir)
     return st
 
@@ -245,13 +155,13 @@ def _read_event_from_traces(stream):
 
 def read_traces():
     """
-    Read trace files
+    Read traces from the files or paths specified in the configuration.
 
     :return: Traces
     :rtype: :class:`obspy.core.stream.Stream`
     """
     logger.info('Reading traces...')
-    stream = _read_trace_files()
+    stream = _read_asdf_traces() + _read_trace_files()
     _read_event_from_traces(stream)
     logger.info('Reading traces: done')
     logger.info('---------------------------------------------------')
